Move concurrent OPTIMIZE test in Iceberg
