+++ plugin/trino-iceberg/src/test/java/io/trino/plugin/iceberg/TestIcebergLocalConcurrentWrites.java
@@ -15,22 +15,28 @@ package io.trino.plugin.iceberg;
 
 import com.google.common.collect.ImmutableList;
 import io.airlift.concurrent.MoreFutures;
+import io.trino.plugin.blackhole.BlackHolePlugin;
 import io.trino.testing.AbstractTestQueryFramework;
+import io.trino.testing.DistributedQueryRunner;
 import io.trino.testing.MaterializedResult;
 import io.trino.testing.QueryRunner;
 import io.trino.testing.sql.TestTable;
 import org.junit.jupiter.api.RepeatedTest;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.Callable;
 import java.util.concurrent.CyclicBarrier;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
+import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import static com.google.common.base.Preconditions.checkState;
+import static com.google.common.base.Verify.verify;
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static io.airlift.concurrent.MoreFutures.tryGetFutureValue;
 import static io.trino.testing.QueryAssertions.getTrinoExceptionCause;
@@ -49,7 +55,10 @@ final class TestIcebergLocalConcurrentWrites
     protected QueryRunner createQueryRunner()
             throws Exception
     {
-        return IcebergQueryRunner.builder().build();
+        DistributedQueryRunner queryRunner = IcebergQueryRunner.builder().build();
+        queryRunner.installPlugin(new BlackHolePlugin());
+        queryRunner.createCatalog("blackhole", "blackhole");
+        return queryRunner;
     }
 
     // Repeat test with invocationCount for better test coverage, since the tested aspect is inherently non-deterministic.
@@ -1041,6 +1050,97 @@ final class TestIcebergLocalConcurrentWrites
         finally {
             executor.shutdownNow();
             assertThat(executor.awaitTermination(10, SECONDS)).isTrue();
+        }
+    }
+
+    @Test
+    public void testOptimizeDuringWriteOperations()
+            throws Exception
+    {
+        runOptimizeDuringWriteOperations(true);
+        runOptimizeDuringWriteOperations(false);
+    }
+
+    private void runOptimizeDuringWriteOperations(boolean useSmallFiles)
+            throws Exception
+    {
+        int threads = 5;
+        int deletionThreads = threads - 1;
+        int rows = 12;
+        int rowsPerThread = rows / deletionThreads;
+
+        CyclicBarrier barrier = new CyclicBarrier(threads);
+        ExecutorService executor = newFixedThreadPool(threads);
+
+        // Slow down the delete operations so optimize is more likely to complete
+        String blackholeTable = "blackhole_table_" + randomNameSuffix();
+        assertUpdate("CREATE TABLE blackhole.default.%s (a INT, b INT) WITH (split_count = 1, pages_per_split = 1, rows_per_page = 1, page_processing_delay = '3s')".formatted(blackholeTable));
+
+        try (TestTable table = newTrinoTable(
+                "test_optimize_during_write_operations",
+                "(int_col INT)")) {
+            String tableName = table.getName();
+
+            // Testing both situations where a file is fully removed by the delete operation and when a row level delete is required.
+            if (useSmallFiles) {
+                for (int i = 0; i < rows; i++) {
+                    assertUpdate(format("INSERT INTO %s VALUES %s", tableName, i), 1);
+                }
+            }
+            else {
+                String values = IntStream.range(0, rows).mapToObj(String::valueOf).collect(Collectors.joining(", "));
+                assertUpdate(format("INSERT INTO %s VALUES %s", tableName, values), rows);
+            }
+
+            List<Future<List<Boolean>>> deletionFutures = IntStream.range(0, deletionThreads)
+                    .mapToObj(threadNumber -> executor.submit(() -> {
+                        barrier.await(10, SECONDS);
+                        List<Boolean> successfulDeletes = new ArrayList<>();
+                        for (int i = 0; i < rowsPerThread; i++) {
+                            try {
+                                int rowNumber = threadNumber * rowsPerThread + i;
+                                getQueryRunner().execute(format("DELETE FROM %s WHERE int_col = %s OR ((SELECT count(*) FROM blackhole.default.%s) > 42)", tableName, rowNumber, blackholeTable));
+                                successfulDeletes.add(true);
+                            }
+                            catch (RuntimeException e) {
+                                successfulDeletes.add(false);
+                            }
+                        }
+                        return successfulDeletes;
+                    }))
+                    .collect(toImmutableList());
+
+            Future<?> optimizeFuture = executor.submit(() -> {
+                try {
+                    barrier.await(10, SECONDS);
+                    // Allow for some deletes to start before running optimize
+                    Thread.sleep(50);
+                    assertUpdate("ALTER TABLE %s EXECUTE optimize".formatted(tableName));
+                }
+                catch (Exception e) {
+                    throw new RuntimeException(e);
+                }
+            });
+
+            List<String> expectedValues = new ArrayList<>();
+            for (int threadNumber = 0; threadNumber < deletionThreads; threadNumber++) {
+                List<Boolean> deleteOutcomes = deletionFutures.get(threadNumber).get();
+                verify(deleteOutcomes.size() == rowsPerThread);
+                for (int rowNumber = 0; rowNumber < rowsPerThread; rowNumber++) {
+                    boolean successfulDelete = deleteOutcomes.get(rowNumber);
+                    if (!successfulDelete) {
+                        expectedValues.add(String.valueOf(threadNumber * rowsPerThread + rowNumber));
+                    }
+                }
+            }
+
+            optimizeFuture.get();
+            assertThat(expectedValues.size()).isGreaterThan(0).isLessThan(rows);
+            assertQuery("SELECT * FROM " + tableName, "VALUES " + String.join(", ", expectedValues));
+        }
+        finally {
+            executor.shutdownNow();
+            executor.awaitTermination(10, SECONDS);
         }
     }
